+++ mathics/builtin/system.py
@@ -469,15 +469,16 @@ class SystemMemory(Predefined):
     """
 
     name = "$SystemMemory"
-    
+
     def evaluate(self, evaluation) -> Integer:
         try:
             import psutil
+
             totalmem = psutil.virtual_memory().total
             return Integer(total)
         except:
             return String(SymbolFailed)
-    
+
 
 class MemoryAvailable(Builtin):
     """
@@ -489,11 +490,12 @@ class MemoryAvailable(Builtin):
     >> MemoryAvailable[]
      = ...
     """
-    
+
     def apply_0(self, evaluation) -> Integer:
         """MemoryAvailable[]"""
         try:
             import psutil
+
             totalmem = psutil.virtual_memory().available
             return Integer(total)
         except:
@@ -510,20 +512,24 @@ class MemoryInUse(Builtin):
     >> MemoryInUse[]
      = ...
     """
-    
+
     def apply_0(self, evaluation) -> Integer:
         """MemoryInUse[]"""
         # Partially borrowed from https://code.activestate.com/recipes/577504/
         from itertools import chain
         from sys import getsizeof
+
         definitions = evaluation.definitions
         seen = set()
         default_size = getsizeof(0)
-        handlers = {tuple: iter,
-                    list: iter,
-                    dict: (lambda d: chain.from_iterable(d.items())),
-                    set: iter,
-                    frozenset: iter,}
+        handlers = {
+            tuple: iter,
+            list: iter,
+            dict: (lambda d: chain.from_iterable(d.items())),
+            set: iter,
+            frozenset: iter,
+        }
+
         def sizeof(obj):
             if id(obj) in seen:
                 return 0
@@ -536,10 +542,6 @@ class MemoryInUse(Builtin):
             return s
 
         return Integer(sizeof(definitions))
-
-                
-            
-        
 
 
 class Share(Builtin):
@@ -555,7 +557,7 @@ class Share(Builtin):
     >> Share[]
      = ...
     """
-    
+
     def apply_0(self, evaluation) -> Integer:
         """Share[]"""
         # TODO: implement a routine that swap all the definitions,
@@ -571,5 +573,3 @@ class Share(Builtin):
         # remplace them by references.
         # Return the amount of memory recovered.
         return Integer(0)
-
-